Split out weight_info.rs in Spot
